// src/app/shared/models/EmployeeADImportedModel.ts
export class EmployeeADImportedModel {

    accountName?: string;
    hrCode?: string;
    firstName?: string;
    lastName?: string;
    displayName?: string;
    email?: string;
    IsExist?: boolean;
    checkbox?: boolean;
    enabled?: boolean;
    serviceAccount: boolean;
    archiveAccount: boolean;
    accountType: string;

  }